# backend/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 1.0.0
  description: Verified B2B Deal Network API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /organizations/{orgId}/documents:
    get:
      operationId: listDocuments
      tags: [documents]
      summary: List documents for an organization
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DocumentRef"
    post:
      operationId: uploadDocument
      tags: [documents]
      summary: Upload and encrypt a document to IPFS via Pinata
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/DocumentUploadInput"
      responses:
        "201":
          description: Uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentRef"
  /documents/{docId}/download:
    get:
      operationId: downloadDocument
      tags: [documents]
      summary: Decrypt and stream a document (authorized access only)
      security:
        - bearerAuth: []
      parameters:
        - name: docId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Document stream
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary

  # ─── Wallets ─────────────────────────────────────────────────────────────────
...
  /connections:
    get:
      operationId: listConnections
      tags: [connections]
      summary: List connections for the current user's organizations
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Connections
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Connection"
    post:
      operationId: requestConnection
      tags: [connections]
      summary: Request a connection to another organization
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionInput"
      responses:
        "201":
          description: Connection request sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /connections/{connectionId}:
    get:
      operationId: getConnection
      tags: [connections]
      summary: Get a connection by ID
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"
  /connections/{connectionId}/respond:
    post:
      operationId: respondToConnection
      tags: [connections]
      summary: Accept, decline, or profile-only a connection request
      security:
        - bearerAuth: []
      parameters:
        - name: connectionId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConnectionRespondInput"
      responses:
        "200":
          description: Updated connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Connection"

  # ─── Workspaces ──────────────────────────────────────────────────────────────
  /workspaces:
    post:
      operationId: createWorkspace
      tags: [workspaces]
      summary: Create a deal workspace from an accepted connection
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WorkspaceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealWorkspace"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentRef:
      type: object
      additionalProperties: false
      required: [id, organizationId, docType, ipfsCid, contentHash, uploadedBy, uploadedAt]
      properties:
        id:
          type: string
        organizationId:
          type: string
        docType:
          type: string
          enum: [REGISTRATION_CERT, ID, PROOF_OF_ADDRESS, AGREEMENT, INVOICE, EVIDENCE, OTHER]
        ipfsCid:
          type: string
        contentHash:
          type: string
        uploadedBy:
          type: string
        uploadedAt:
          type: string
          format: date-time
    DocumentUploadInput:
      type: object
      additionalProperties: false
      required: [file, docType]
      properties:
        file:
          type: string
          description: Binary file content (multipart/form-data)
        docType:
          type: string
          enum: [REGISTRATION_CERT, ID, PROOF_OF_ADDRESS, AGREEMENT, INVOICE, EVIDENCE, OTHER]

    # ── Wallets ─────────────────────────────────────────────────────────────────
...
    Connection:
      type: object
      additionalProperties: false
      required: [id, requestingOrgId, targetOrgId, status, createdAt]
      properties:
        id:
          type: string
        requestingOrgId:
          type: string
        targetOrgId:
          type: string
        reason:
          type: ["string", "null"]
        proposedDealCategory:
          type: ["string", "null"]
        approxValue:
          type: ["string", "null"]
        timeline:
          type: ["string", "null"]
        requiresNda:
          type: boolean
        status:
          type: string
          enum: [PENDING, ACCEPTED, DECLINED, PROFILE_ONLY]
        createdAt:
          type: string
          format: date-time
        respondedAt:
          type: ["string", "null"]
          format: date-time
    ConnectionInput:
      type: object
      additionalProperties: false
      required: [requestingOrgId, targetOrgId]
      properties:
        requestingOrgId:
          type: string
        targetOrgId:
          type: string
        reason:
          type: string
        proposedDealCategory:
          type: string
        approxValue:
          type: string
        timeline:
          type: string
        requiresNda:
          type: boolean
    ConnectionRespondInput:
      type: object
      additionalProperties: false
      required: [response]
      properties:
        response:
          type: string
          enum: [ACCEPTED, DECLINED, PROFILE_ONLY]
        createWorkspace:
          type: boolean

    # ── Workspaces ───────────────────────────────────────────────────────────────
    DealWorkspace:
      type: object
      additionalProperties: false
      required: [id, title, status, participantOrgIds, createdAt]
      properties:
        id:
          type: string
        connectionId:
          type: ["string", "null"]
        title:
          type: string
        objectiveSummary:
          type: ["string", "null"]
        status:
          type: string
          enum:
            [
              DISCOVERED,
              CONNECTION_REQUESTED,
              QUALIFYING,
              WORKSPACE_ACTIVE,
              REQUIREMENTS_DEFINED,
              PROPOSAL_SUBMITTED,
              NEGOTIATING,
              TERMS_AGREED,
              INTERNAL_APPROVAL,
              CONTRACT_SIGNED,
              AWAITING_FUNDING,
              FUNDED,
              IN_PROGRESS,
              AWAITING_ACCEPTANCE,
              COMPLETED,
              CANCELLED,
              REVISION_REQUIRED,
              DISPUTED,
              REFUNDED,
              PARTIALLY_SETTLED,
            ]
        participantOrgIds:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceInput:
      type: object
      additionalProperties: false
      required: [connectionId, title]
      properties:
        connectionId:
          type: string
        title:
          type: string
        objectiveSummary:
          type: string
...